requirements.txt
# IBM ELM MCP Server Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=1.0.0
PyMuPDF>=1.24.0
matplotlib>=3.8.0
markdown>=3.5
